chore(code-garden): extract content scheduler validation [2026-W29]
Non-functional change from repo audit 2026-W29.
Finding: Content Scheduler route inlines all validation helpers.
