+++ docs/repo-audits/2026-W29.md
@@ -111,7 +111,7 @@ AI agents
 - *Why it matters:* The file has grown since W27 (952→972) with the markdown-checkbox feature. Each addition makes it harder to navigate and increases the risk of accidental coupling between unrelated UI concerns. The pattern set by `tasks/_*` extraction should propagate here.
 - *Severity:* Medium (carryover from W27/W28)
 
-**[Low] Content Scheduler route `contentScheduler.ts` (374 lines) inlines all validation helpers.**
+**[Low] Content Scheduler route `contentScheduler.ts` (374 lines) inlines all validation helpers.** ✅ [PR #251](https://github.com/Stoffer-Industries/sindustries/pull/251)
 - *Where:* `services/tasks-api/src/routes/contentScheduler.ts`
 - *Why it matters:* Validation helpers (`parseId`, `parseDate`, `validateBody`) are duplicated from the pattern established in `tasks/_validation.ts`. This is acceptable at current size but inconsistent with the rest of the codebase convention.
 - *Severity:* Low
