+++ services/tasks-api/src/routes/contentScheduler.ts
@@ -25,36 +25,20 @@ import {
   getXClient,
   TERMINAL_STATUSES
 } from './contentSchedulerPublish.ts';
+import {
+  parseDate,
+  parseId,
+  uuidPattern,
+  validateBody,
+  validSources,
+  validStatuses
+} from './contentSchedulerValidation.ts';
 
-const MAX_BODY = 1000;
-
-const uuidPattern = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
-
-const validSources = new Set(['ops_notes', 'cto_craft', 'manual', 'other']);
-const validStatuses = new Set(['draft', 'queued', 'approved', 'published', 'removed']);
 
 function actor(req: any): string {
   const header = req.headers['x-actor'];
   if (typeof header === 'string' && header.trim().length > 0) return header.trim();
   return 'unknown';
-}
-
-function parseId(raw: string): string | null {
-  return uuidPattern.test(raw) ? raw : null;
-}
-
-function parseDate(value: unknown): Date | null | 'invalid' {
-  if (value === undefined || value === null || value === '') return null;
-  const d = new Date(value as string);
-  return Number.isNaN(d.valueOf()) ? 'invalid' : d;
-}
-
-function validateBody(body: unknown): string | null {
-  if (typeof body !== 'string') return 'body must be a string';
-  const trimmed = body.trim();
-  if (trimmed.length === 0) return 'body must not be empty';
-  if (body.length > MAX_BODY) return `body must be <= ${MAX_BODY} characters`;
-  return null;
 }
 
 export const contentSchedulerRouter = Router();
